refactor(home): 탭 설정을 constants.ts로 분리 (#22)
