+++ src/app/(main)/home/_ui/constants.ts
@@ -1,0 +1,14 @@
+import { LockIcon } from '@/shared/assets/icons'
+import type { HomeTabConfig } from './HomeTabs'
+
+const MY_HOME_TABS: HomeTabConfig[] = [
+  { id: 'posts', label: '게시글' },
+  { id: 'breeders', label: '즐겨찾는 브리더', Icon: LockIcon },
+]
+
+const BREEDER_HOME_TABS: HomeTabConfig[] = [
+  { id: 'listings', label: '분양목록' },
+  { id: 'posts', label: '게시글' },
+]
+
+export { MY_HOME_TABS, BREEDER_HOME_TABS }
